+++ common/web/types/src/util/index.ts
@@ -1,0 +1,2 @@
+export * from './consts.js';
+export * from './util.js';
