+++ common/web/types/src/util/util.ts
@@ -1,3 +1,4 @@
+import { MATCH_HEX_ESCAPE, MATCH_QUAD_ESCAPE } from './consts.js';
 /**
  * xml2js will not place single-entry objects into arrays. Easiest way to fix
  * this is to box them ourselves as needed. Ensures that o.x is an array.
@@ -15,18 +16,6 @@ export function boxXmlArray(o: any, x: string): void {
     }
   }
 }
-
-// TODO-LDML: #7569 the below regex works, but captures more than it should
-// (it would include \u{fffffffffffffffff } which
-// is overlong and has a space at the end.) The second regex does not work yet.
-export const MATCH_HEX_ESCAPE = /\\u{([0-9a-fA-F ]{1,})}/g;
-// const MATCH_HEX_ESCAPE = /\\u{((?:(?:[0-9a-fA-F]{1,5})|(?:10[0-9a-fA-F]{4})(?: (?!}))?)+)}/g;
-
-/** regex for single quad escape such as \u0127 or \U00000000 */
-export const CONTAINS_QUAD_ESCAPE = /(?:\\u([0-9a-fA-F]{4})|\\U([0-9a-fA-F]{8}))/;
-
-/** regex for single quad escape such as \u0127 */
-export const MATCH_QUAD_ESCAPE = new RegExp(CONTAINS_QUAD_ESCAPE, 'g');
 
 export class UnescapeError extends Error {
 }
